Add JV880 background

+++ src/backend/lcd.h
@@ -62,22 +62,22 @@ public:
 struct lcd_t {
     mcu_t* mcu = nullptr;
 
-    size_t width = 0;
+    size_t width  = 0;
     size_t height = 0;
 
     uint32_t color1 = 0;
     uint32_t color2 = 0;
 
     // all the variables in this group are updated by the MCU via LCD_Write
-    uint32_t LCD_DL = 0, LCD_N = 0, LCD_F = 0, LCD_D = 0, LCD_C = 0, LCD_B = 0, LCD_ID = 0, LCD_S = 0;
-    uint32_t LCD_DD_RAM = 0, LCD_AC = 0, LCD_CG_RAM = 0;
+    uint32_t LCD_DL       = 0, LCD_N  = 0, LCD_F      = 0, LCD_D = 0, LCD_C = 0, LCD_B = 0, LCD_ID = 0, LCD_S = 0;
+    uint32_t LCD_DD_RAM   = 0, LCD_AC = 0, LCD_CG_RAM = 0;
     uint32_t LCD_RAM_MODE = 0;
     uint8_t  LCD_Data[80]{};
     uint8_t  LCD_CG[64]{};
 
-    std::atomic<uint8_t> enable = 0;
-    std::atomic<uint8_t> button_enable = 0;
-    uint8_t              contrast = 8;
+    std::atomic<uint8_t>  enable        = 0;
+    std::atomic<uint32_t> button_enable = 0;
+    uint8_t              contrast       = 8;
 
     uint32_t buffer[lcd_height_max][lcd_width_max]{};
 
@@ -94,6 +94,6 @@ bool LCD_Start(lcd_t& lcd);
 void LCD_Stop(lcd_t& lcd);
 void LCD_Write(lcd_t& lcd, uint32_t address, uint8_t data);
 void LCD_Enable(lcd_t& lcd, uint32_t enable);
-void LCD_ButtonEnable(lcd_t& lcd, uint8_t enable);
+void LCD_ButtonEnable(lcd_t& lcd, uint32_t enable);
 void LCD_SetContrast(lcd_t& lcd, uint8_t contrast);
 void LCD_Render(lcd_t& lcd);
